Moved blender download last workfile, and workfile is up to date check implementation. Using blender self.report, changing ops prefixes, changing enum short names, added sync_server availability check.
